requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.12
dj-rest-auth==2.1.11
Django==3.2.6
django-braces==1.14.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-oauth-toolkit==1.5.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.1
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
jwcrypto==1.0
oauthlib==3.1.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
wrapt==1.12.1
### IF HEROKU DEPLOYMENT
####django-heroku==0.3.1
